requirements.txt
amqp==2.5.1
asgi-redis==1.4.2
asgiref==1.1.2
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
billiard==3.5.0.5
celery==4.1.1
certifi==2019.9.11
cffi==1.12.3
channels==1.1.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==1.3.0
dj-static==0.0.6
Django==2.2
django-admin-rangefilter==0.3.12
django-celery-beat==1.1.1
django-celery-results==1.1.0
django-cors-headers==2.1.0
django-extensions==2.2.1
django-rest-auth==0.9.5
django-rest-recaptcha==1.0.2
django-rest-swagger==2.2.0
djangorestframework==3.9.3
djangorestframework-gis==0.14
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
kombu==4.2.0
MarkupSafe==1.1.1
msgpack-python==0.5.6
openapi-codec==1.3.2
Pillow==6.0.0
postgis==1.0.4
psycopg2==2.7.7
pycparser==2.19
PyHamcrest==1.9.0
pystache==0.5.4
python-dateutil==2.8.0
python-http-client==3.2.1
pytz==2018.9
redis==2.10.5
requests==2.22.0
sendgrid==3.6.5
sendgrid-django==4.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
static==1.1.1
static3==0.7.0
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.15
vine==1.3.0
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.6.0
